+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0]
 ### Added
-- New Command `alias:add` to add a new alias.
+- New command `alias:add` to add a new alias.
+- New command `alias:remove` to remove a configured alias.
+- New command `aliases:list` to show a list of all configured aliases.
 
 ## [1.0.0] - 2018-03-13
 ### Added